+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Enhancements:
       ``rtmidi.midiconstants``.
 
 Examples:
+    * Added ``midiclock`` example to show how to receive MIDI timimng clock.
     * Changed ``midioutwrapper`` example to use
       ``rtmidi.midiutil.open_midiout`` so that it also works on Windows (with
       no virtual ports support).